refactor: :recycle: rename file to validation_errors_to_check_errors
